resample: Calculate full sinc table (e.g. no sinc interpolation)
On many systems taking couple of kilo-bytes for full sinc table is not
a problem. With a straight filter table the number of MAC operations
needed drops to one quarter. Also the inner loop of the filter becomes
more simple and easier to optimize. So the full sinc table version
runs much faster in most cases. The trade off is some kilo-bytes
bigger memory usage and more CPU needed when initializing the
resampler.

configure.ac: Add --enable-resample-full-sinc-table conf flag

Use this flag to enable resampling with full sinc table.