Tidy the speex/*.h includes some more
This is a partial revert of 785e61a593c61ebc0afee5d87eca9cc73e16d6a6.

The _BUILD_SPEEX special case isn't needed in the public headers, and
we don't need '../include' relative paths in the internal files.

Don't include speex.h in jitter.c, it's not needed there, and no longer
present in this source tree.

Switch from speex_types.h to speexdsp_types.h in arch.h, it's not clear
yet what the best thing to do with the latter will be (since they are
duplicates of each other and must always stay synchronised), but the
former is no longer present in this source tree anymore either.