resample: complete speex_resampler_reset_mem
sample_frac_num is the array most likely to be non-zero after resampling a
previous stream, but the other arrays can also be left at non-zero values in
other situations.

Using a previous sample_frac_num can mean that the first input sample is
ignored.  This may usually not be significant, but may just be detectable when
only one channel of a stereo signal is affected.

Reviewed by Jean-Marc Valin <jmvalin@mozilla.com> at
https://bugzilla.mozilla.org/show_bug.cgi?id=973374

Signed-off-by: Tristan Matthews <le.businessman@gmail.com>